@@ -10,9 +10,9 @@
   },
   "dependencies": {
     "framer-motion": "^11.9.0",
-    "next": "13.5.7",
-    "react": "^18",
-    "react-dom": "^18"
+    "next": "^15.0.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@types/node": "^20",
