@@ -1,6 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {}
+const nextConfig = {};
 
-module.exports = {
-  swcMinify: false,
-};
+export default nextConfig;
